fixing missing bloc builders
The home_page.dart's source code get some style refact.

Impelemnted the state listener for the bloc on the page/view side.

Because of showModalBottomSheet(...) need 2 bloc provider and bloc builder.
Because of this function there are 2 hiden page in 1 page class.

Therfore the next step is: need upgrade the repository layer.